scrollprize.org: build index for browser client-side from minified metadata.json
